bench done

+++ CalculatorF/Calculator.fs
@@ -8,7 +8,7 @@ module Calculator  =
         | Divide = 3
         | Multiply = 4
 
-    let Calculate (val1:decimal, operation:Operation, val2:decimal) =
+    let Calculate (val1:int32, operation:Operation, val2:int32) =
         match operation with
         | Operation.Plus -> Some (val1 + val2)
         | Operation.Minus -> Some (val1 - val2)
